角色删除时检测有没有绑定角色

+++ fushuai-dmanager/src/main/java/com/manager/web/RoleController.java
@@ -14,9 +14,13 @@ import com.kh.pojo.ResponseResult;
 import com.kh.pojo.entity.MenuInfo;
 import com.kh.pojo.entity.RoleInfo;
 
+import com.kh.pojo.entity.RoleMenuInfo;
+import com.kh.pojo.entity.UserRoleInfo;
 import com.kh.utils.TwitterIdWorker;
 import com.manager.dao.MenuDao;
+import com.manager.service.RoleMenuService;
 import com.manager.service.RoleService;
+import com.manager.service.UserRoleService;
 import com.manager.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -47,13 +51,14 @@ public class RoleController {
     private RedisTemplate<String, String> redisTemplate;
 
     @Autowired
-    private UserService userService;
-    @Autowired
     private RoleService roleService;
+    @Autowired
+    private UserRoleService userRoleService;
+    @Autowired
+    private RoleMenuService roleMenuService;
 
     @Autowired
     private MenuDao menuDao;
-
 
 
     /**
@@ -72,37 +77,67 @@ public class RoleController {
     //角色删除
     @RequestMapping("deleteRoleById")
     @ResponseBody
-    public int deleteRoleById(@RequestBody Map<String, String> map) {
-        Long id = Long.valueOf(map.get("id"));
-        int i = roleService.deleteRoleById(id);
-        return i;
+    public ResponseResult deleteRoleById(@RequestBody RoleInfo map) {
+        ResponseResult responseResult = new ResponseResult();
+        Long id = map.getId();
+        //检测有没有绑定用户
+        UserRoleInfo userRoleInfo = userRoleService.selectUserRoleByRoleId(id);
+        System.out.println(userRoleInfo + "=========================================roleMenuInfos");
+
+        if (userRoleInfo != null) {
+            responseResult.setCode(500);
+            responseResult.setError("绑定有角色");
+        } else {
+            try {
+                //删除角色权限
+                roleMenuService.deleteByRoleId(id);
+                //删除角色
+                roleService.deleteRoleById(id);
+
+                responseResult.setCode(200);
+                responseResult.setSuccess("删除成功");
+            } catch (Exception e) {
+                responseResult.setCode(500);
+                responseResult.setError("删除失败");
+            }
+        }
+        return responseResult;
     }
 
     //角色添加
     @RequestMapping("addRole")
     @ResponseBody
-    public int addRole(@RequestBody RoleInfo r) {
+    public ResponseResult addRole(@RequestBody RoleInfo r) {
+        ResponseResult responseResult = new ResponseResult();
 
         //设置id为雪花id
         TwitterIdWorker twitterIdWorker = new TwitterIdWorker();
         long id = twitterIdWorker.nextId();
         r.setId(id);
+
         //添加
-        int i = roleService.addRole(r);
-        return i;
+        try {
+            roleService.addRole(r);
+            responseResult.setCode(200);
+            responseResult.setSuccess("添加成功");
+        } catch (Exception e) {
+            responseResult.setCode(500);
+            responseResult.setError("添加失败");
+        }
+        return responseResult;
     }
 
-//权限回显
+    //权限回显
     @RequestMapping("/findMenu")
     @ResponseBody
-    public List<MenuInfo> findMenu(){
+    public List<MenuInfo> findMenu() {
 
         return getForMenuInfo(0L);
     }
 
-    public List<MenuInfo> getForMenuInfo(Long mid){
+    public List<MenuInfo> getForMenuInfo(Long mid) {
         List<MenuInfo> firstMenuInfo = menuDao.findByParentId(mid);
-        if(firstMenuInfo!=null){
+        if (firstMenuInfo != null) {
             for (MenuInfo menuInfo : firstMenuInfo) {
                 menuInfo.setMenuInfoList(getForMenuInfo(menuInfo.getId()));
             }
@@ -111,45 +146,45 @@ public class RoleController {
     }
 
 
-
-
     @Autowired
     JdbcTemplate jdbcTemplate;
-//角色绑定权限
+
+    //角色绑定权限
     @RequestMapping("/addRm")
     @Transactional
     @ResponseBody
-    public ResponseResult addRm(@RequestBody RoleInfo roleInfo){
-        ResponseResult responseResult=new ResponseResult();
+    public ResponseResult addRm(@RequestBody RoleInfo roleInfo) {
+        ResponseResult responseResult = new ResponseResult();
 
         try {
             roleService.addRole(roleInfo);
-            String sql1="delete from base_role_menu where roleId=?";
-            jdbcTemplate.update(sql1,roleInfo.getId());
+
+            String sql1 = "delete from base_role_menu where roleId=?";
+            jdbcTemplate.update(sql1, roleInfo.getId());
 
             String[] ids = roleInfo.getIds();
-            if(ids!=null&&!ids.equals("")) {
+            if (ids != null && !ids.equals("")) {
                 //中间表添加权限
-                String sql="insert into base_role_menu(id,roleId,menuId) values(?,?,?)";
-                List<Object[]>list=new ArrayList<>();
+                String sql = "insert into base_role_menu(id,roleId,menuId) values(?,?,?)";
+                List<Object[]> list = new ArrayList<>();
                 TwitterIdWorker twitterIdWorker = new TwitterIdWorker();
                 for (String s : ids) {
 
-                    list.add(new Object[]{twitterIdWorker.nextId(),roleInfo.getId(),Long.parseLong(s)});
+                    list.add(new Object[]{twitterIdWorker.nextId(), roleInfo.getId(), Long.parseLong(s)});
                 }
-                jdbcTemplate.batchUpdate(sql,list);
+                jdbcTemplate.batchUpdate(sql, list);
             }
 
             responseResult.setCode(200);
         } catch (NumberFormatException e) {
-            e.printStackTrace();
+
+            responseResult.setCode(500);
+        } catch (Exception e) {
             responseResult.setCode(500);
         }
 
         return responseResult;
     }
-
-
 
 
 }
